update: optimize folder
