+++ src/pages/Content/VideoType/lynda.ts
@@ -2,9 +2,9 @@
   Lynda video
  */
 
-import { getItem } from './localStorage';
+import { getItem } from '../modules/localStorage';
 // @ts-ignore
-import {hiddenSubtitleCssInject,dealSubtitle} from './utils.ts'
+import {hiddenSubtitleCssInject,dealSubtitle} from '../modules/utils.ts'
 
 const sub = {
   pre: '',
