+++ webpack.config.js
@@ -38,6 +38,7 @@ var options = {
     options: path.join(__dirname, 'src', 'pages', 'Options', 'index.jsx'),
     popup: path.join(__dirname, 'src', 'pages', 'Popup', 'index.jsx'),
     background: path.join(__dirname, 'src', 'pages', 'Background', 'index.js'),
+    devtools: path.join(__dirname, 'src', 'pages', 'Devtools', 'index.js'),
     contentScript: path.join(__dirname, 'src', 'pages', 'Content', 'index.ts'),
   },
   chromeExtensionBoilerplate: {
@@ -167,6 +168,17 @@ var options = {
       ),
       filename: 'background.html',
       chunks: ['background'],
+    }),
+    new HtmlWebpackPlugin({
+      template: path.join(
+        __dirname,
+        'src',
+        'pages',
+        'Devtools',
+        'index.html',
+      ),
+      filename: 'devtools.html',
+      chunks: ['devtools'],
     }),
     new WriteFilePlugin(),
   ],
